fix: require NUT-09 restore before committing added mint
Self-review of #156: swallowing restore failures after addMint left a
tracked mint with permanent zero balance — the original bug mode. Make
restore required before the mint list commit on Android and iOS, load
transactions after recovery, and drop the no-op best-effort helper.
